Extracted the streaming FDR paths out of PercolatorFdr
* ComputeStreamingCompetitionQvalues, ComputeFullPopulationPrecursorFdrStreaming
  and StreamingFirstPassQ moved to StreamingFdr; 433 lines verified verbatim
* StreamingFdr is public, not internal: Osprey.Tasks calls it across an assembly
  boundary that InternalsVisibleTo does not cover
* PercolatorFdr.cs 1,269 -> 838 lines; Stellar regression byte-identical

See ai/todos/active/TODO-20260727_osprey_percolatorfdr_decomposition.md
